refactor: restructure profile tabs and add styles
