+++ src/screens/private/client/Profile/Tabs/ExportCard/index.ts
@@ -1,0 +1,1 @@
+export { ExportCard } from './ExportCard';
