+++ src/screens/private/client/Profile/Tabs/FavoritosTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FavoritosTab';
