+++ src/screens/private/client/Profile/Tabs/HistoricoCompras/index.ts
@@ -1,0 +1,1 @@
+export { default } from './HistoricoCompras';
