+++ src/screens/private/client/Profile/Tabs/MenuNavegacao/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MenuNavegacao';
