+++ src/screens/private/client/Profile/Tabs/MeusAgendamentos/MeusAgendamentos.tsx
@@ -1,11 +1,5 @@
 import React, { useEffect, useMemo, useState } from 'react';
-import {
-  Text,
-  View,
-  StyleSheet,
-  ScrollView,
-  useWindowDimensions,
-} from 'react-native';
+import { Text, View, ScrollView, useWindowDimensions } from 'react-native';
 import { useAppointmentStore } from '@stores/Appointment';
 import { useFavoriteStore } from '@stores/Favorite';
 import { useColors } from '@theme/ThemeProvider';
@@ -14,6 +8,7 @@ import { RateServiceModal } from '@components/features/RateServiceModal';
 import { Appointment } from '@stores/Appointment/types';
 import { AppointmentCard } from '@components/features/AppointmentCard';
 import { FontAwesome } from '@expo/vector-icons';
+import { createStyles } from './styles';
 
 function MeusAgendamentos() {
   const { appointments, fetchAppointments } = useAppointmentStore();
@@ -75,7 +70,7 @@ function MeusAgendamentos() {
       <FontAwesome
         name="calendar-o"
         size={40}
-        color={colors.secondaryBeige}
+        color={colors.textTertiary}
         style={{ marginBottom: 12 }}
       />
       <Text style={styles.emptyText}>{text}</Text>
@@ -94,7 +89,7 @@ function MeusAgendamentos() {
           <FontAwesome
             name="clock-o"
             size={18}
-            color="white"
+            color={colors.primaryWhite}
             style={{ marginRight: 8 }}
           />
           <Text style={styles.sectionTitle}>Próximos Agendamentos</Text>
@@ -129,12 +124,12 @@ function MeusAgendamentos() {
         <View
           style={[
             styles.sectionHeader,
-            { backgroundColor: colors.primaryGreen },
+            { backgroundColor: colors.successText },
           ]}>
           <FontAwesome
             name="check-circle-o"
             size={18}
-            color="white"
+            color={colors.primaryWhite}
             style={{ marginRight: 8 }}
           />
           <Text style={styles.sectionTitle}>Histórico</Text>
@@ -190,58 +185,5 @@ function MeusAgendamentos() {
     </ScrollView>
   );
 }
-
-const createStyles = (colors: any) =>
-  StyleSheet.create({
-    container: {
-      flex: 1,
-    },
-    pageTitle: {
-      fontSize: 28,
-      fontFamily: 'Afacad-Bold',
-      color: colors.primaryBlack,
-      marginBottom: 24,
-    },
-    section: {
-      marginBottom: 32,
-    },
-    sectionHeader: {
-      backgroundColor: colors.primaryBlue,
-      paddingVertical: 12,
-      paddingHorizontal: 16,
-      borderRadius: 8,
-      marginBottom: 16,
-      flexDirection: 'row',
-      alignItems: 'center',
-    },
-    sectionTitle: {
-      fontSize: 16,
-      fontFamily: 'Afacad-Bold',
-      color: colors.primaryWhite,
-    },
-    grid: {
-      flexDirection: 'row',
-      flexWrap: 'wrap',
-      gap: 16,
-    },
-    gridItem: {
-      width: '100%',
-    },
-    emptyContainer: {
-      padding: 32,
-      alignItems: 'center',
-      justifyContent: 'center',
-      backgroundColor: 'white',
-      borderRadius: 12,
-      borderWidth: 1,
-      borderColor: '#F0F0F0',
-      borderStyle: 'dashed',
-    },
-    emptyText: {
-      fontSize: 16,
-      fontFamily: 'Afacad-Regular',
-      color: colors.textTertiary,
-    },
-  });
 
 export default MeusAgendamentos;
