+++ src/screens/private/client/Profile/Tabs/MeusAgendamentos/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MeusAgendamentos';
