+++ src/screens/private/client/Profile/Tabs/NotificacoesContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NotificacoesContent';
