+++ src/screens/private/client/Profile/Tabs/PagamentosTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PagamentosTab';
